First work
Tried not to touch any of the UI stuff, but I did make a textView for the date. Set up MainViewModel.java - going off of Labs 3 and 4. Hit a roadblock because I haven't done Lab 5 yet, which is super important. Used the starter code's SimpleSubject instead of Subject for now because it's identical to what we've used in labs; be cautious of that because if we switch to Subject, we'll have to change names all over the place. I don't think MainActivity works rn because my computer can't handle it, but most of the groundwork is error-free. Stayed away from fragments too for now.